+++ src/components/SavedBeats.js
@@ -1,24 +1,22 @@
 import React, { useState } from "react";
 import BeatButton from "./BeatButton";
+import EdmPads from "./EdmPads";
 import Tracks from "./Tracks";
 import Visuals from "./Visuals";
+import EdmPad from "./EdmPads";
 
 function SavedBeats({ sounds }) {
-  // console.log(sounds)
+
   const [playingSongs, setPlayingSongs] = useState([])
   const [globalMute, setGlobalMute] = useState(false)
 
-  const edmSounds = sounds.filter(sound => sound.genre === "EDM")
+  // const edmSounds = sounds.filter(sound => sound.genre === "EDM")
   const wubstepSounds = sounds.filter(sound => sound.genre === "Wubstep")
   const trapSounds = sounds.filter(sound => sound.genre === "Trap")
 
-  // console.log(edmSounds)
-  // console.log(wubstepSounds)
-  // console.log(trapSounds)
-
-  const edmElements = edmSounds.map(sound => {
-    return (<BeatButton key={sound.id} sound={sound} />)
-  })
+  // const edmElements = edmSounds.map(sound => {
+  //   return (<BeatButton key={sound.id} sound={sound} />)
+  // })
   const wubstepElements = wubstepSounds.map(sound => {
     return (<BeatButton key={sound.id} sound={sound} />)
   })
@@ -26,19 +24,19 @@ function SavedBeats({ sounds }) {
     return (<BeatButton key={sound.id} sound={sound} />)
   })
 
-  function playEdmSong() {
-    console.log("play song")
-    //fetch all edmSounds from the database
-    fetch(`http://localhost:8003/sounds`)
-      .then(r => r.json())
-      .then(beats => {
-        const edmBeats = beats.filter(beat => (beat.genre === "EDM"))
-        const edmSong = edmBeats.filter(beat => (beat.isMuted === false))
-        setPlayingSongs(edmSong)
-        setGlobalMute(false)
-      })
+  // function playEdmSong() {
+  //   console.log("play song")
+  //   //fetch all edmSounds from the database
+  //   fetch(`http://localhost:8003/sounds`)
+  //     .then(r => r.json())
+  //     .then(beats => {
+  //       const edmBeats = beats.filter(beat => (beat.genre === "EDM"))
+  //       const edmSong = edmBeats.filter(beat => (beat.isMuted === false))
+  //       setPlayingSongs(edmSong)
+  //       setGlobalMute(false)
+  //     })
     //if they are not muted, then make a player for each of them which starts upon them all being loaded
-  }
+  // }
   function playWubstepSong() {
     console.log("play song")
     fetch(`http://localhost:8003/sounds`)
@@ -50,6 +48,7 @@ function SavedBeats({ sounds }) {
         setGlobalMute(false)
       })
   }
+
   function playTrapSong() {
     console.log("play song")
     fetch(`http://localhost:8003/sounds`)
@@ -70,21 +69,13 @@ function SavedBeats({ sounds }) {
     console.log(beat.ref)
     return (globalMute ? '' : <Tracks key={beat.id} src={beat.ref} />)
   })
-  console.log(globalMute)
-
-  console.log(audioSources)
 
   return (
     <div>
       <div className="mixer">
         <Visuals />
         <div className="padscontainer">
-          <h1>EDM Sounds:</h1>
-          <div className="button-container">
-            {edmElements}
-            <br></br>
-            <button className="play-btn" onClick={playEdmSong}>Play EDM Song</button>
-          </div>
+          <EdmPads sounds={sounds} setGlobalMute={setGlobalMute} setPlayingSongs={setPlayingSongs} />
           <h1>Wubstep Sounds:</h1>
           <div className="button-container">
             {wubstepElements}
